requirements.txt
click==8.0.1
requests==2.26.0
kucoin-python==1.0.9